refactor: MainActivity 내 DB CRUD 메서드를 OmokDAO로 분리
